temp

+++ aku/training/train_run_clm.py
@@ -471,8 +471,7 @@ def main():
         model = MistralForCausalLM.from_pretrained(pretrained_model_name_or_path=None, 
                                                     config=config, 
                                                     state_dict=OrderedDict(),
-                                                    use_flash_attention_2=True,
-                                                    torch_dtype=torch.bfloat16)
+                                                    use_flash_attention_2=True)
 
         print("mistral config:",config)
         print("mistral model architecture:",model)
